requirements.txt
aiofiles==23.2.1
anyio==4.4.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
easyocr==1.7.1
etelemetry==0.3.1
filelock==3.14.0
fitz==0.0.1.dev2
Flask==3.0.3
Flask-Cors==4.0.1
frontend==0.0.3
fsspec==2024.5.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.3.0
prov==2.0.0
pyclipper==1.3.0.post5
pydot==2.0.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.5.22
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traits==6.3.2
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.0
Werkzeug==3.0.3
